refactor: コンポーネントを分割し、ロジックをカスタムフックに移動してコードを整理
